requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
PySocks==1.7.1
requests==2.31.0